+++ lib/puppet/provider/cisco_interface/nxapi.rb
@@ -43,7 +43,6 @@ Puppet::Type.type(:cisco_interface).provide(:nxapi) do
     :switchport_mode,
     :vrf,
     :access_vlan,
-    :channel_group,
     :description,
     :encapsulation_dot1q,
     :ipv4_address,